+++ .gitignore
@@ -1,2 +1,3 @@
 .vscode/
+.cursor/
 .DS_store
