+++ src/InterfaceGeneration/Parsing/ASTParser/ASTParser.jl
@@ -1,0 +1,7 @@
+module ASTParser
+
+export ASTParserImpl, parse
+
+include("ASTParserImpl.jl")
+
+end
