+++ src/InterfaceGeneration/Parsing/Parser.jl
@@ -1,0 +1,2 @@
+abstract type Parser end
+abstract type AbstractModule end
