+++ src/InterfaceGeneration/Parsing/Parsing.jl
@@ -1,0 +1,15 @@
+module Parsing
+
+export Parser, parse
+
+include("Parser.jl")
+include("RegexParser/RegexParser.jl")
+include("ASTParser/ASTParser.jl")
+
+using .RegexParser
+using .ASTParser
+
+import .RegexParser: parse
+import .ASTParser: parse
+
+end
