+++ src/InterfaceGeneration/Parsing/RegexParser/RegexParser.jl
@@ -1,0 +1,7 @@
+module RegexParser
+
+export RegexParserImpl, parse
+
+include("RegexParserImpl.jl")
+
+end
